Client: code compatibility refactoring for Windows and Linux
