+++ client/packages/vita_virtual_device/src/virtual_utils.rs
@@ -1,7 +1,6 @@
 use crate::virtual_button::{Button, DpadDirection};
-use crate::virtual_config::{TouchConfig, TriggerConfig};
-use crate::virtual_touch::{Point, TouchAction};
-use vita_reports::{ButtonsData, TouchReport};
+use crate::virtual_config::TriggerConfig;
+use vita_reports::ButtonsData;
 
 /// Computes the D-Pad direction based on the button states.
 pub fn compute_dpad_direction(buttons: &ButtonsData) -> DpadDirection {
@@ -15,21 +14,6 @@ pub fn compute_dpad_direction(buttons: &ButtonsData) -> DpadDirection {
         (false, false, true, false) => DpadDirection::West,
         (false, false, false, true) => DpadDirection::East,
         _ => DpadDirection::None,
-    }
-}
-
-/// Converts DpadDirection to axis values suitable for uinput.
-pub fn dpad_direction_to_axis_values(direction: DpadDirection) -> (i32, i32) {
-    match direction {
-        DpadDirection::North => (0, -1),
-        DpadDirection::NorthEast => (1, -1),
-        DpadDirection::East => (1, 0),
-        DpadDirection::SouthEast => (1, 1),
-        DpadDirection::South => (0, 1),
-        DpadDirection::SouthWest => (-1, 1),
-        DpadDirection::West => (-1, 0),
-        DpadDirection::NorthWest => (-1, -1),
-        DpadDirection::None => (0, 0),
     }
 }
 
@@ -63,22 +47,4 @@ pub fn get_pressed_buttons(
         .into_iter()
         .filter_map(|(pressed, button)| if pressed { Some(button) } else { None })
         .collect()
-}
-
-/// Processes touch reports and returns a list of touch actions.
-pub fn process_touch_reports(
-    touch_reports: &[TouchReport],
-    touch_config: &Option<TouchConfig>,
-) -> Vec<TouchAction> {
-    let mut actions = Vec::new();
-    if let Some(TouchConfig::Zones(zones)) = touch_config {
-        for touch in touch_reports {
-            if let Some(zone) = zones.locate_at_point(&Point(touch.x.into(), touch.y.into())) {
-                if let Some(action) = zone.action {
-                    actions.push(action);
-                }
-            }
-        }
-    }
-    actions
 }
